Remove rails/init.rb, Fixup railtie.rb  Fixes #450
railtie.rb:
rubber needs to be initialized after the rails configurations are
loaded. The right place is initializer block.

init.rb: This file was last touched 4 years ago and does not fit in
rails 4.1.x framework. The init code belongs in Railtie#initializer
function, which is already coded (2 years ago). So this file was
redundant at best.